refactor: move web filters/interceptor to no.entur.mummu.web; cut customizer complexity
Rename the no.entur.mummu.logging package to no.entur.mummu.web — it now holds
request-limiting (MaxCountInterceptor/MaxCountProperties) as well as observability
(LargeResponseLoggingFilter, LogSanitizer), so 'logging' no longer fit. Also flatten
countParameterDescriptionCustomizer into a stream + extracted helper (Sonar cognitive
complexity 27 -> under 15).
