+++ src/test/java/no/entur/mummu/web/MaxCountInterceptorIntegrationTest.java
@@ -1,4 +1,4 @@
-package no.entur.mummu.logging;
+package no.entur.mummu.web;
 
 import no.entur.mummu.MummuApplication;
 import org.junit.jupiter.api.Test;
